requirements.txt
langchain
streamlit
python-dotenv
langchain_community
langchain-openai
chromadb
pypdf